Move utils/pytest into testutils
This reduces the number of directories that cause issues if imported
into production code.
