+++ src/sentry/testutils/pytest/template/credentials.json
