--- src/sentry/utils/pytest/__init__.py
@@ -1,10 +1,0 @@
-pytest_plugins = [
-    "sentry.utils.pytest.sentry",
-    "sentry.utils.pytest.selenium",
-    "sentry.utils.pytest.fixtures",
-    "sentry.utils.pytest.unittest",
-    "sentry.utils.pytest.kafka",
-    "sentry.utils.pytest.relay",
-    "sentry.utils.pytest.metrics",
-    "sentry.utils.pytest.stale_database_reads",
-]
